Move interfaces from models to correct types folder
